REFACT: move corda util file
