+++ tests/test_initialization.py
@@ -57,10 +57,10 @@ from peft import (
     set_peft_model_state_dict,
 )
 from peft.tuners.lora.config import CordaConfig, CordaInitConfig
+from peft.tuners.lora.corda import preprocess_corda
 from peft.tuners.lora.layer import LoraLayer
 from peft.utils import infer_device
 from peft.utils.constants import PEFT_TYPE_TO_PREFIX_MAPPING
-from peft.utils.corda_utils import preprocess_corda
 from peft.utils.hotswap import hotswap_adapter
